# api/swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of the Spin-It API.",
    "title": "Spin-It API",
    "contact": {
      "name": "Rick Roche",
      "url": "https://www.rickroche.com",
      "email": "a@b.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "basePath": "/",
  "paths": {
...
    "/v1/playlists": {
      "post": {
        "description": "Create a new playlist on Spotify using the matched tracks from the match endpoint",
        "summary": "Create a playlist on Spotify using the matched tracks",
        "operationId": "playlists",
        "parameters": [
          {
            "description": "The playlist to be created",
            "name": "createPlaylist",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreatePlaylistRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Match response",
            "schema": {
              "$ref": "#/definitions/CreatePlaylistResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Album": {
      "type": "object",
      "title": "Album ...",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "spin-it/internal/models"
    },
    "Artist": {
      "type": "object",
      "title": "Artist ...",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "spin-it/internal/models"
    },
...
    "CreatePlaylistRequest": {
      "description": "Create a new playlist",
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "name": {
          "description": "in: body",
          "type": "string",
          "x-go-name": "Name"
        },
        "public": {
          "type": "boolean",
          "x-go-name": "Public"
        },
        "tracks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Track"
          },
          "x-go-name": "Tracks"
        }
      },
      "x-go-package": "spin-it/api"
    },
    "CreatePlaylistResponse": {
      "description": "Create a new playlist response",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "public": {
          "type": "boolean",
          "x-go-name": "IsPublic"
        },
        "totalTracks": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TotalTracks"
        }
      },
      "x-go-package": "spin-it/api"
    },
...
    "Track": {
      "type": "object",
      "title": "Track ...",
      "properties": {
        "album": {
          "$ref": "#/definitions/Album"
        },
        "artists": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Artist"
          },
          "x-go-name": "Artists"
        },
        "confidence": {
          "type": "string",
          "x-go-name": "Confidence"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "playcount": {
          "type": "string",
          "x-go-name": "PlayCount"
        },
        "rank": {
          "type": "string",
          "x-go-name": "Rank"
        },
        "searchedFor": {
          "type": "string",
          "x-go-name": "SearchedFor"
        }
      },
      "x-go-package": "spin-it/internal/models"
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    },
    "oauth2": {
      "type": "oauth2",
      "in": "header",
      "flow": "accessCode",
      "authorizationUrl": "/oauth2/auth",
      "tokenUrl": "/oauth2/token",
      "scopes": {
        "bar": "foo"
      }
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}